feat: executado comando php artisan migrate

+++ app/database/migrations/2021_04_18_232401_create_locations_table.php
@@ -21,7 +21,7 @@ class CreateLocationsTable extends Migration
                 latitude e longitude para um determinado país
                 já vou deixar ele configurado para ser chave estrangeira
             */
-            $table->integer('country_id')->unsigned();
+            $table->BigInteger('country_id')->unsigned();
             $table->foreign('country_id')->references('id')->on('countries')->onDelete('cascade');
             $table->integer('latitude');
             $table->integer('longitude');
